Done with repositories

+++ src/main/java/com/danil/crud/repository/GsonLabelRepositoryImpl.java
@@ -6,8 +6,8 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.lang.reflect.Type;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 
 import com.danil.crud.model.Label;
@@ -16,50 +16,61 @@ import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
 public class GsonLabelRepositoryImpl implements LabelRepository {
+    private static int maxID = 0;
     private final Gson gson = new Gson();
     private final File file;
 
     public GsonLabelRepositoryImpl(String filename) {
         this.file = new File(filename);
+        setMaxId();
     }
 
+    @Override
     public void save(Label label) {
         if (label == null) {
             throw new NullPointerException("Tried to save null label");
         }
 
         HashMap<Integer, Label> collection = getAll();
+        if (collection == null) {
+            collection = new HashMap<>();
+        }
         if (!collection.containsKey(label.getId())) {
             collection.put(label.getId(), label);
             saveAll(collection);
+            setMaxId();
         }
     }
 
+    @Override
     public void update(Label label) {
         if (label == null) {
             throw new NullPointerException("Tried to update null label");
         }
 
         HashMap<Integer, Label> collection = getAll();
-        Label labelToUpdate = collection.get(label.getId());
-        if (labelToUpdate != null && labelToUpdate.getStatus() != LabelStatus.DELETED) {
+        if (collection == null) {
+            return;
+        }
+        if (collection.containsKey(label.getId())) {
             collection.put(label.getId(), label);
             saveAll(collection);
         }
     }
 
+    @Override
     public HashMap<Integer, Label> getAll() {
         String contents = null;
         try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
             contents = reader.readLine();
         } catch (IOException e) {
             System.err.println("Cannot read " + file.getName() + ": " + e.getMessage());
-            return new HashMap<>();
+            return null;
         }
 
-        Type token = new TypeToken<ArrayList<Label>>() {
-        }.getType();
-        ArrayList<Label> labelList = gson.fromJson(contents, token);
+        TypeToken<ArrayList<Label>> collectionType = new TypeToken<ArrayList<Label>>() {
+        };
+        ArrayList<Label> labelList = gson.fromJson(contents, collectionType);
         HashMap<Integer, Label> result = new HashMap<>();
         if (labelList == null) {
             return result;
@@ -70,7 +81,45 @@ public class GsonLabelRepositoryImpl implements LabelRepository {
         return result;
     }
 
-    private void saveAll(HashMap<Integer, Label> collection) {
+    @Override
+    public HashMap<Integer, Label> getAllExceptDeleted() {
+        String contents = null;
+        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
+            contents = reader.readLine();
+        } catch (IOException e) {
+            System.err.println("Cannot read " + file.getName() + ": " + e.getMessage());
+            return null;
+        }
+
+        TypeToken<ArrayList<Label>> collectionType = new TypeToken<ArrayList<Label>>() {
+        };
+        ArrayList<Label> labelList = gson.fromJson(contents, collectionType);
+        HashMap<Integer, Label> result = new HashMap<>();
+        if (labelList == null) {
+            return result;
+        }
+        for (Label label : labelList) {
+            if (label.getStatus() != LabelStatus.DELETED) {
+                result.put(label.getId(), label);
+            }
+        }
+        return result;
+    }
+
+    @Override
+    public Label getById(Integer id) {
+        if (id == null) {
+            throw new NullPointerException("Tried to get label with null id");
+        }
+        HashMap<Integer, Label> collection = getAll();
+        if (collection == null) {
+            return null;
+        }
+        return collection.get(id);
+    }
+
+    @Override
+    public void saveAll(HashMap<Integer, Label> collection) {
         if (collection == null) {
             throw new NullPointerException("Tried to save null collection");
         }
@@ -84,17 +133,39 @@ public class GsonLabelRepositoryImpl implements LabelRepository {
         }
     }
 
+    @Override
     public void deleteById(Integer id) {
         if (id == null) {
-            throw new NullPointerException("Tried to delete element by null id");
+            throw new NullPointerException("Tried to delete label by null id");
         }
 
         HashMap<Integer, Label> collection = getAll();
+        if (collection == null) {
+            return;
+        }
         Label target = collection.get(id);
         if (target != null) {
             target.delete();
             collection.put(id, target);
             saveAll(collection);
+        }
+    }
+
+    @Override
+    public int getMaxId() {
+        return maxID;
+    }
+
+    private void setMaxId() {
+        HashMap<Integer, Label> collection = getAll();
+        if (collection == null) {
+            return;
+        }
+        Collection<Label> values = collection.values();
+        for (Label val : values) {
+            if (maxID <= val.getId()) {
+                maxID = val.getId() + 1;
+            }
         }
     }
 }
